Added PDF functionality and fixed color boundaries
I added the functionality to convert the first page of a pdf into an image. I also fixed some bugs regarding color boundaries. It turns out, I was using the BGR boundaries instead of using HSV boundaries, which is why the colors overlapped with each other. I also came to understand that there should be two regions for red boundaries, one for ultraviolet and one for infared.